% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cu_personnel}
\alias{cu_personnel}
\title{University of Colorado Personnel Data}
\format{A tibble with 25,314 rows and 8 variables:
\describe{
  \item{campus}{CU campus, a factor}
  \item{roster.id}{a unique ID}
  \item{dept.group}{Department group}
  \item{dept}{department, a factor}
  \item{job.family}{job family, a factor}
  \item{job.title}{job title, a factor}
  \item{percent}{percent}
  \item{funding}{funding} 
}}
\source{
\url{https://www.cu.edu/doc/fy18personnelcurosterallcampusespdf}
}
\usage{
cu_personnel
}
\description{
University of Colorado Personnel Data
}
\keyword{datasets}
