% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cu_personnel}
\alias{cu_personnel}
\title{University of Colorado Personnel Data}
\format{A tibble with 23,458 rows and 7 variables:
\describe{
  \item{campus}{CU campus, a factor}
  \item{roster.id}{a unique ID}
  \item{job.family}{job family, a factor}
  \item{job.title}{job title, a factor}
  \item{dept}{department, a factor}
  \item{percent}{percent }
  \item{funding}{funding} 
}}
\source{
\url{http://www.cu.edu/doc/fy17-personnel-cu-roster-all-campuses.pdf}
}
\usage{
cu_personnel
}
\description{
University of Colorado Personnel Data
}
\keyword{datasets}
